requirements.txt
Django>=4.2
django-ninja
djangorestframework-simplejwt
psycopg2-binary
pydantic
